lab/01-model-deploy/mlpipelines/pipelines/requirements.txt
boto3
sagemaker
PyYAML